+++ internal/acceptance/bake/README.md
@@ -1,0 +1,8 @@
+# Acceptance Tests
+There is not much about the structure of these tests to document.  They're
+vanilla Gomega test suites, with a normal amount of quirk (like files named
+"test" that don't test anything).
+
+## Current Tests
+### bake_test.go
+Considers log output as part of the API but otherwise is cromulent.
