+++ internal/acceptance/bake/fixtures/bosh-vars/variable-1.yml
