+++ internal/acceptance/bake/fixtures/icon
