+++ internal/acceptance/bake/fixtures/instance-groups/some-instance-group.yml
