+++ internal/acceptance/bake/fixtures/instance-groups2/some-other-instance-group.yml
