+++ internal/acceptance/bake/fixtures/jobs/some-job.yml
