+++ internal/acceptance/bake/fixtures/jobs2/some-other-job.yml
