+++ internal/acceptance/bake/fixtures/metadata-with-multiple-stemcells.yml
