+++ internal/acceptance/bake/fixtures/metadata-with-stemcell-criteria.yml
