+++ internal/acceptance/bake/fixtures/metadata-with-stemcell-tarball.yml
