+++ internal/acceptance/bake/fixtures/multiple-stemcells/stemcell-1.tgz
