+++ internal/acceptance/bake/fixtures/multiple-stemcells/stemcell-2.tgz
