+++ internal/acceptance/bake/fixtures/properties/some-templated-property.yml
