+++ internal/acceptance/bake/fixtures/releases/not-a-tarball.txt
