+++ internal/acceptance/bake/fixtures/runtime-config/some-runtime-config.yml
