+++ internal/acceptance/bake/fixtures/single-stemcell/stemcell.tgz
