+++ internal/acceptance/bake/fixtures/variables-file
